refactor: split death.ts into death/ directory
